requirements.txt
Django
djangorestframework
pandas
gunicorn
django-cors-headers 